reorganized Modal component style directory
