+++ src/components/Helpers/Modal/style/CloseModalButtonStyle.js
